resolve extension metadata uniformly during object creation
