+++ src/project/project-context.ts
@@ -107,6 +107,32 @@ import { createTempContext } from "../core/temp.ts";
 
 import { onCleanup } from "../core/cleanup.ts";
 import { once } from "../core/once.ts";
+import { Zod } from "../resources/types/zod/schema-types.ts";
+
+const mergeExtensionMetadata = async (
+  context: ProjectContext,
+  pOptions: RenderOptions,
+) => {
+  // this will mutate context.config.project to merge
+  // in any project metadata from extensions
+  if (context.config) {
+    const extensions = await pOptions.services.extension.extensions(
+      undefined,
+      context.config,
+      context.isSingleFile ? undefined : context.dir,
+      { builtIn: false },
+    );
+    const projectMetadata = extensions.filter((extension) =>
+      extension.contributes.metadata?.project
+    ).map((extension) => {
+      return Zod.ProjectConfig.parse(extension.contributes.metadata!.project);
+    });
+    context.config.project = mergeProjectMetadata(
+      context.config.project,
+      ...projectMetadata,
+    );
+  }
+};
 
 export async function projectContext(
   path: string,
@@ -143,6 +169,16 @@ export async function projectContext(
       cachedEnv = await computeProjectEnvironment(notebookContext, project);
       return cachedEnv;
     }
+  };
+
+  const returnResult = async (
+    context: ProjectContext,
+  ) => {
+    if (renderOptions) {
+      await mergeExtensionMetadata(context, renderOptions);
+    }
+    onCleanup(context.cleanup);
+    return context;
   };
 
   while (true) {
@@ -362,8 +398,7 @@ export async function projectContext(
           config: configFiles,
           configResources: projectConfigResources(dir, projectConfig, type),
         };
-        onCleanup(result.cleanup);
-        return result;
+        return await returnResult(result);
       } else {
         debug(`projectContext: Found Quarto project in ${dir}`);
         const temp = createTempContext({
@@ -431,8 +466,7 @@ export async function projectContext(
           config: configFiles,
           configResources: projectConfigResources(dir, projectConfig),
         };
-        onCleanup(result.cleanup);
-        return result;
+        return await returnResult(result);
       }
     } else {
       const nextDir = dirname(dir);
@@ -511,8 +545,7 @@ export async function projectContext(
             context.files.input = [input];
           }
           debug(`projectContext: Found Quarto project in ${originalDir}`);
-          onCleanup(context.cleanup);
-          return context;
+          return await returnResult(context);
         } else {
           return undefined;
         }
